sonar when

+++ src/creatures/creature.h
@@ -195,7 +195,7 @@ class Creature : virtual public Thing
 			}
 		}
 
-		void setBaseSpeed(uint32_t newBaseSpeed) {
+		void setBaseSpeed(uint16_t newBaseSpeed) {
 			baseSpeed = newBaseSpeed;
 		}
 		uint16_t getBaseSpeed() const {
@@ -563,7 +563,7 @@ class Creature : virtual public Thing
 		uint32_t blockCount = 0;
 		uint32_t blockTicks = 0;
 		uint32_t lastStepCost = 1;
-		uint32_t baseSpeed = 110;
+		uint16_t baseSpeed = 110;
 		uint32_t mana = 0;
 		int32_t varSpeed = 0;
 		int32_t health = 1000;
